refactor: 폴더명 컨벤션 맞추기
